vecstore: move vector store implementation tests into a common test
Previously, the in-memory and persistent vector stores maintained their
own disjoint tests. This patch moves those tests that test common
functionality into a single joint test that is referenced from both
implementation unit tests.

Epic: CRDB-42943
