+++ pkg/sql/vecindex/vecstore/in_memory_store_test.go
@@ -28,14 +28,6 @@ func TestInMemoryStore(t *testing.T) {
 
 	ctx := internal.WithWorkspace(context.Background(), &internal.Workspace{})
 
-	childKey2 := ChildKey{PartitionKey: 2}
-	childKey10 := ChildKey{PartitionKey: 10}
-	childKey20 := ChildKey{PartitionKey: 20}
-	childKey30 := ChildKey{PartitionKey: 30}
-	childKey40 := ChildKey{PartitionKey: 40}
-	childKey50 := ChildKey{PartitionKey: 50}
-	childKey60 := ChildKey{PartitionKey: 60}
-
 	store := NewInMemoryStore(2, 42)
 	quantizer := quantize.NewUnQuantizer(2)
 
@@ -72,174 +64,7 @@ func TestInMemoryStore(t *testing.T) {
 		}, vectors)
 	})
 
-	t.Run("search empty root partition", func(t *testing.T) {
-		txn := beginTransaction(ctx, t, store)
-		defer commitTransaction(ctx, t, store, txn)
-
-		searchSet := SearchSet{MaxResults: 2}
-		partitionCounts := []int{0}
-		level, err := txn.SearchPartitions(
-			ctx, []PartitionKey{RootKey}, vector.T{1, 1}, &searchSet, partitionCounts)
-		require.NoError(t, err)
-		require.Equal(t, LeafLevel, level)
-		require.Nil(t, searchSet.PopResults())
-		require.Equal(t, 0, partitionCounts[0])
-	})
-
-	t.Run("add to root partition", func(t *testing.T) {
-		txn := beginTransaction(ctx, t, store)
-		defer commitTransaction(ctx, t, store, txn)
-
-		// Add to root partition.
-		count, err := txn.AddToPartition(ctx, RootKey, vector.T{1, 2}, childKey10)
-		require.NoError(t, err)
-		require.Equal(t, 1, count)
-		count, err = txn.AddToPartition(ctx, RootKey, vector.T{7, 4}, childKey20)
-		require.NoError(t, err)
-		require.Equal(t, 2, count)
-		count, err = txn.AddToPartition(ctx, RootKey, vector.T{4, 3}, childKey30)
-		require.NoError(t, err)
-		require.Equal(t, 3, count)
-
-		// Add duplicate and expect value to be overwritten
-		count, err = txn.AddToPartition(ctx, RootKey, vector.T{5, 5}, childKey30)
-		require.NoError(t, err)
-		require.Equal(t, 3, count)
-
-		// Search root partition.
-		searchSet := SearchSet{MaxResults: 2}
-		partitionCounts := []int{0}
-		level, err := txn.SearchPartitions(
-			ctx, []PartitionKey{RootKey}, vector.T{1, 1}, &searchSet, partitionCounts)
-		require.NoError(t, err)
-		require.Equal(t, Level(1), level)
-		result1 := SearchResult{QuerySquaredDistance: 1, ErrorBound: 0, CentroidDistance: 2.2361, ParentPartitionKey: 1, ChildKey: childKey10}
-		result2 := SearchResult{QuerySquaredDistance: 32, ErrorBound: 0, CentroidDistance: 7.0711, ParentPartitionKey: 1, ChildKey: childKey30}
-		results := searchSet.PopResults()
-		roundResults(results, 4)
-		require.Equal(t, SearchResults{result1, result2}, results)
-		require.Equal(t, 3, partitionCounts[0])
-	})
-
-	var root *Partition
-	t.Run("get root partition", func(t *testing.T) {
-		txn := beginTransaction(ctx, t, store)
-		defer commitTransaction(ctx, t, store, txn)
-
-		// Get root partition.
-		var err error
-		root, err = txn.GetPartition(ctx, RootKey)
-		require.NoError(t, err)
-		require.Equal(t, Level(1), root.Level())
-		require.Equal(t, []ChildKey{childKey10, childKey20, childKey30}, root.ChildKeys())
-		require.Equal(t, vector.T{0, 0}, root.Centroid())
-
-		// Get partition centroid + full vectors.
-		results := []VectorWithKey{
-			{Key: ChildKey{PartitionKey: RootKey}},
-			{Key: ChildKey{PrimaryKey: PrimaryKey{11}}},
-			{Key: ChildKey{PrimaryKey: PrimaryKey{0}}},
-		}
-		err = txn.GetFullVectors(ctx, results)
-		require.NoError(t, err)
-		require.Equal(t, vector.T{0, 0}, results[0].Vector)
-		require.Equal(t, vector.T{100, 200}, results[1].Vector)
-		require.Nil(t, results[2].Vector)
-	})
-
-	t.Run("replace root partition", func(t *testing.T) {
-		txn := beginTransaction(ctx, t, store)
-		defer commitTransaction(ctx, t, store, txn)
-
-		// Replace root partition.
-		_, err := txn.GetPartition(ctx, RootKey)
-		require.NoError(t, err)
-		vectors := vector.T{4, 3}.AsSet()
-		quantizedSet := quantizer.Quantize(ctx, &vectors)
-		newRoot := NewPartition(quantizer, quantizedSet, []ChildKey{childKey2}, 2)
-		require.NoError(t, txn.SetRootPartition(ctx, newRoot))
-		newRoot, err = txn.GetPartition(ctx, RootKey)
-		require.NoError(t, err)
-		require.Equal(t, Level(2), newRoot.Level())
-		require.Equal(t, []ChildKey{childKey2}, newRoot.ChildKeys())
-
-		searchSet := SearchSet{MaxResults: 2}
-		partitionCounts := []int{0}
-		level, err := txn.SearchPartitions(
-			ctx, []PartitionKey{RootKey}, vector.T{2, 2}, &searchSet, partitionCounts)
-		require.NoError(t, err)
-		require.Equal(t, Level(2), level)
-		result3 := SearchResult{QuerySquaredDistance: 5, ErrorBound: 0, CentroidDistance: 0, ParentPartitionKey: 1, ChildKey: childKey2}
-		require.Equal(t, SearchResults{result3}, searchSet.PopResults())
-		require.Equal(t, 1, partitionCounts[0])
-	})
-
-	var partitionKey1 PartitionKey
-	t.Run("insert another partition and update it", func(t *testing.T) {
-		txn := beginTransaction(ctx, t, store)
-		defer commitTransaction(ctx, t, store, txn)
-
-		_, err := txn.GetPartition(ctx, RootKey)
-		require.NoError(t, err)
-		partitionKey1, err = txn.InsertPartition(ctx, root)
-		require.NoError(t, err)
-		require.Equal(t, PartitionKey(2), partitionKey1)
-		count, err := txn.RemoveFromPartition(ctx, partitionKey1, childKey20)
-		require.NoError(t, err)
-		require.Equal(t, 2, count)
-
-		// Try to remove the same key again.
-		count, err = txn.RemoveFromPartition(ctx, partitionKey1, childKey20)
-		require.NoError(t, err)
-		require.Equal(t, 2, count)
-
-		// Add an alternate element and add duplicate, expecting value to be overwritten.
-		count, err = txn.AddToPartition(ctx, partitionKey1, vector.T{-1, 0}, childKey40)
-		require.NoError(t, err)
-		require.Equal(t, 3, count)
-		count, err = txn.AddToPartition(ctx, partitionKey1, vector.T{1, 1}, childKey40)
-		require.NoError(t, err)
-		require.Equal(t, 3, count)
-
-		searchSet := SearchSet{MaxResults: 2}
-		partitionCounts := []int{0}
-		level, err := txn.SearchPartitions(
-			ctx, []PartitionKey{partitionKey1}, vector.T{1, 1}, &searchSet, partitionCounts)
-		require.NoError(t, err)
-		require.Equal(t, Level(1), level)
-		result4 := SearchResult{QuerySquaredDistance: 0, ErrorBound: 0, CentroidDistance: 1.4142, ParentPartitionKey: 2, ChildKey: childKey40}
-		result5 := SearchResult{QuerySquaredDistance: 1, ErrorBound: 0, CentroidDistance: 2.2361, ParentPartitionKey: 2, ChildKey: childKey10}
-		require.Equal(t, SearchResults{result4, result5}, roundResults(searchSet.PopResults(), 4))
-		require.Equal(t, 3, partitionCounts[0])
-	})
-
-	t.Run("search multiple partitions at leaf level", func(t *testing.T) {
-		txn := beginTransaction(ctx, t, store)
-		defer commitTransaction(ctx, t, store, txn)
-
-		_, err := txn.GetPartition(ctx, RootKey)
-		require.NoError(t, err)
-
-		vectors := vector.MakeSet(2)
-		vectors.Add(vector.T{4, -1})
-		vectors.Add(vector.T{2, 8})
-		quantizedSet := quantizer.Quantize(ctx, &vectors)
-		partition := NewPartition(quantizer, quantizedSet, []ChildKey{childKey50, childKey60}, LeafLevel)
-		partitionKey2, err := txn.InsertPartition(ctx, partition)
-		require.NoError(t, err)
-		require.Equal(t, PartitionKey(3), partitionKey2)
-
-		searchSet := SearchSet{MaxResults: 2}
-		partitionCounts := []int{0, 0}
-		level, err := txn.SearchPartitions(
-			ctx, []PartitionKey{partitionKey1, partitionKey2}, vector.T{3, 1}, &searchSet, partitionCounts)
-		require.NoError(t, err)
-		require.Equal(t, Level(1), level)
-		result4 := SearchResult{QuerySquaredDistance: 4, ErrorBound: 0, CentroidDistance: 1.41, ParentPartitionKey: 2, ChildKey: childKey40}
-		result5 := SearchResult{QuerySquaredDistance: 5, ErrorBound: 0, CentroidDistance: 2.24, ParentPartitionKey: 2, ChildKey: childKey10}
-		require.Equal(t, SearchResults{result4, result5}, roundResults(searchSet.PopResults(), 2))
-		require.Equal(t, []int{3, 2}, partitionCounts)
-	})
+	commonStoreTests(ctx, t, store, quantizer)
 
 	t.Run("delete full vector", func(t *testing.T) {
 		txn := beginTransaction(ctx, t, store)
